Improved tq query, options, safer stdout reading, updated README

+++ tractor-purge.py
@@ -8,29 +8,48 @@ import re
 import shutil
 import datetime
 import logging
-
+from optparse import OptionParser
 
 
 
 ####################################
-# Settings
+# Option parser and constants
 
-TQ = '/opt/pixar/Tractor-2.2/bin/tq'
-TRACTOR_ENGINE = 'TRACTOR-ENGINE' # You can use an IP address here
-CMD_LOGS_DIR = '/var/spool/tractor/cmd-logs'
-PURGE_LOG = '/var/tmp/tractor-purge.log'
-DAYS = '30' # keep logs and jobs of this age
+parser = OptionParser()
+parser.add_option('-t', '--tq', dest='tq',
+                default='/opt/pixar/Tractor-2.2/bin/tq',
+                help='Absolute path to tq [default: %default]')      
+parser.add_option('-c', '--cmdlogsdir', dest='cmdlogsdir',
+                default='/var/spool/tractor/cmd-logs',
+                help='Absolute path to cmd-logs dir [default: %default]')
+parser.add_option('-l', '--log', dest='logfile',
+                default='/var/tmp/tractor-purge.log',
+                help='Absolute path to tractor-purge log file [default: %default]')
+parser.add_option('-d', '--days', dest='days', default='30',
+                help='Number of days worth of jobs/logs to keep [default: %default]')
+parser.add_option('--deletejobs', action='store_true', dest='deletejobs',
+                default=False,
+                help='Delete jobs from psql database after log deletion [default: %default]')
+parser.add_option('--dryrun', action='store_true', dest='dryrun',
+                default=False,
+                help='Do not perform actual deletion, instead just preview \
+                      deletions [default: %default]')
+
+(options, args) = parser.parse_args()
+
+TQ = options.tq
+CMD_LOGS_DIR = options.cmdlogsdir
+PURGE_LOG = options.logfile
+DAYS = options.days
+DELETE_JOBS = options.deletejobs
+DRY_RUN = options.dryrun
 
 
 ####################################
-# Set up
-
-# TRACTOR_ENGINE environment variable
-if not os.environ.get('TRACTOR_ENGINE') == None:
-    os.environ['TRACTOR_ENGINE'] = TRACTOR_ENGINE
+# General setup
 
 # Logging
-logger = logging.getLogger('Tractor 2.2 jobs/logs purger')
+logger = logging.getLogger('Tractor 2.2 purger')
 hdlr = logging.FileHandler( PURGE_LOG )
 formatter = logging.Formatter('%(asctime)s %(levelname)s %(message)s')
 hdlr.setFormatter(formatter)
@@ -47,25 +66,29 @@ logger.addHandler(ch)
 ####################################
 # Functions
 
-
-def jobs_to_delete(days=None):
+def jobs_to_delete(days):
     """Create list of all jids (equivalient of all jobs to be deleted)
     """
     jids = []
-    command = [TQ, 'jobs', 'not active and spooltime < -' + days + \
-               'd', '--nh', '-c', 'jid']
+    command = [TQ, 'jobs',
+               'not active and not ready and spooltime  < -' + days + 'd',
+               '--noheader', '--archives', '-c', 'jid']
     p = subprocess.Popen( command, stdout=subprocess.PIPE,
                                    stderr=subprocess.STDOUT )
-    for line in iter(p.stdout.readline, b''):
-        sys.stdout.flush()
-        jid = line.rstrip()
-        jids.append(int(jid))
-        logger.info('Added job for deletion: ' + jid)
-        
+    
+    try:
+        for line in iter(p.stdout.readline, b''):
+            sys.stdout.flush()
+            jid = line.rstrip()
+            jids.append(int(jid))
+            logger.info('Added job for deletion: ' + jid)
+    except:
+        logger.warning('Failed to read stdout.')
+
     return jids
 
 
-def get_all_job_folders(cmd_logs_dir=None):
+def get_all_job_folders(cmd_logs_dir):
     """Create list of all job folders
     """
     job_folders = []
@@ -79,7 +102,7 @@ def get_all_job_folders(cmd_logs_dir=None):
     return job_folders
 
 
-def get_job_deletion_list(job_folders=None, jids=None):
+def get_job_deletion_list(job_folders, jids):
     """Compare job folders list against jids list, create deletion list
     """
     delete_list = []
@@ -97,28 +120,38 @@ def get_job_deletion_list(job_folders=None, jids=None):
     return delete_list
 
 
-def delete_logs(delete_list=None):
+def delete_logs(delete_list):
     """Delete the actual log folders
     """
     for job_folder in delete_list:
-        delete = True
-
-        if delete:
+        if not DRY_RUN:
             logger.info( 'Deleting ' + job_folder )
             shutil.rmtree( job_folder )
+        else:
+            logger.info( 'Dry run: (not) deleting ' + job_folder )
 
 
-def delete_tractor_jobs(days=None):
-    """Delete jobs from Tractor (requires that DBArchiving is False in 
-        tractor.config)
+def delete_tractor_jobs(days):
+    """Delete jobs from Tractor. You can also delete jobs manually using:
+    tractor-dbctl --purge-archive-to-year-month YY-MM
     """
-    command = [TQ, '--force', '--yes', 'delete', 'not active and spooltime < -' + \
-               days + 'd']
+    if not DRY_RUN:
+        logger.info( 'Executing tq command to delete jobs...' )
+        command = [TQ, '--force', '--yes', 'delete',
+                   'not active and not ready and spooltime  < -' + days + 'd']
+    else:
+        logger.info( 'Executing tq command to (not) delete jobs...' )
+        command = [TQ, 'jobs', '--archives',
+                   'not active and not ready and spooltime  < -' + days + 'd']
+
     p = subprocess.Popen( command, stdout=subprocess.PIPE,
                                    stderr=subprocess.STDOUT )
-    for line in iter(p.stdout.readline, b''):
-        sys.stdout.flush()
-        logger.info( line.rstrip() )
+    try:
+        for line in iter(p.stdout.readline, b''):
+            sys.stdout.flush()
+            logger.info( line.rstrip() )
+    except:
+        logger.warning('Failed reading stdout.')
 
 
 
@@ -128,7 +161,10 @@ def delete_tractor_jobs(days=None):
 
 if __name__ == '__main__':
 
-    logger.info( 'Tractor purge initiated.' )
+    if not DRY_RUN:
+        logger.info( 'Tractor purge initiated.' )
+    else:
+        logger.info( 'Tractor purge initiated in "dry run" mode.' )
 
     # Queries
     jids = jobs_to_delete(days=DAYS)
@@ -148,10 +184,10 @@ if __name__ == '__main__':
         logger.info( 'No logs to delete.' )
 
     # Delete jobs
-    if len(delete_list) > 0:
-        logger.info( 'Executing tq command to delete jobs...' )
-        delete_tractor_jobs(days=DAYS)
-    else:
-        logger.info( 'No jobs to delete.' )
+    if DELETE_JOBS:
+        if len(delete_list) > 0:
+            delete_tractor_jobs(days=DAYS)
+        else:
+            logger.info( 'No jobs to delete.' )
 
     logger.info( 'Tractor purge done.\n' )
